+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## Unreleased
+### Add
+- add transaction management methods to [DbClient](src/Utility/DbClient.php) and [DbClientInterface](src/Utility/DbClientInterface.php)
+
 ## [v8.0.1] - 2026.02.15
 ### add
 - add support for `doctrine/doctrine-bundle` version 3.0